+++ handlers.go
@@ -1,25 +1,15 @@
 package main
 
 import (
-	"bytes"
-	"context"
 	"encoding/json"
 	"errors"
 	"html/template"
-	"image"
-	"image/png"
 	"io/fs"
 	"io/ioutil"
 	"log"
 	"net/http"
-	"strconv"
-	"strings"
 
 	"github.com/StevenWeathers/thunderdome-planning-poker/pkg/database"
-	"github.com/anthonynsimon/bild/transform"
-	"github.com/gorilla/mux"
-	"github.com/ipsn/go-adorable"
-	"github.com/o1egl/govatar"
 	"github.com/spf13/viper"
 	"gopkg.in/go-playground/validator.v9"
 )
@@ -70,7 +60,7 @@ func ValidateUserPassword(pwd1 string, pwd2 string) (UpdatedPassword string, val
 }
 
 // respondWithJSON takes a payload and writes the response
-func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
+func (s *server) respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
 	response, _ := json.Marshal(payload)
 
 	w.Header().Set("Content-Type", "application/json")
@@ -78,8 +68,8 @@ func respondWithJSON(w http.ResponseWriter, code int, payload interface{}) {
 	w.Write(response)
 }
 
-// gets a JSON request body broken into a key/value map
-func getJSONRequestBody(r *http.Request, w http.ResponseWriter) map[string]interface{} {
+// getJSONRequestBody gets a JSON request body broken into a key/value map
+func (s *server) getJSONRequestBody(r *http.Request, w http.ResponseWriter) map[string]interface{} {
 	body, _ := ioutil.ReadAll(r.Body) // check for errors
 	keyVal := make(map[string]interface{})
 	jsonErr := json.Unmarshal(body, &keyVal) // check for errors
@@ -166,84 +156,6 @@ func (s *server) validateUserCookie(w http.ResponseWriter, r *http.Request) (str
 }
 
 /*
-	Middlewares
-*/
-
-// adminOnly middleware checks if the user is an admin, otherwise reject their request
-func (s *server) adminOnly(h http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		apiKey := r.Header.Get(apiKeyHeaderName)
-		apiKey = strings.TrimSpace(apiKey)
-		var UserID string
-
-		if apiKey != "" {
-			var apiKeyErr error
-			UserID, apiKeyErr = s.database.ValidateAPIKey(apiKey)
-			if apiKeyErr != nil {
-				log.Println("error validating api key : " + apiKeyErr.Error() + "\n")
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-		} else {
-			var cookieErr error
-			UserID, cookieErr = s.validateUserCookie(w, r)
-			if cookieErr != nil {
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-		}
-
-		adminErr := s.database.ConfirmAdmin(UserID)
-		if adminErr != nil {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		ctx := context.WithValue(r.Context(), contextKeyUserID, UserID)
-
-		h(w, r.WithContext(ctx))
-	}
-}
-
-// userOnly validates that the request was made by a valid user
-func (s *server) userOnly(h http.HandlerFunc) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		apiKey := r.Header.Get(apiKeyHeaderName)
-		apiKey = strings.TrimSpace(apiKey)
-		var UserID string
-
-		if apiKey != "" {
-			var apiKeyErr error
-			UserID, apiKeyErr = s.database.ValidateAPIKey(apiKey)
-			if apiKeyErr != nil {
-				log.Println("error validating api key : " + apiKeyErr.Error() + "\n")
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-		} else {
-			var cookieErr error
-			UserID, cookieErr = s.validateUserCookie(w, r)
-			if cookieErr != nil {
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-		}
-
-		_, UserErr := s.database.GetUser(UserID)
-		if UserErr != nil {
-			log.Println("error finding user : " + UserErr.Error() + "\n")
-			s.clearUserCookies(w)
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		ctx := context.WithValue(r.Context(), contextKeyUserID, UserID)
-
-		h(w, r.WithContext(ctx))
-	}
-}
-
-/*
 	Handlers
 */
 
@@ -320,459 +232,6 @@ func (s *server) handleIndex() http.HandlerFunc {
 }
 
 /*
-	Auth Handlers
-*/
-
-// handleLogin attempts to login the user by comparing email/password to whats in DB
-func (s *server) handleLogin() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-		UserEmail := keyVal["warriorEmail"].(string)
-		UserPassword := keyVal["warriorPassword"].(string)
-
-		authedUser, err := s.authUserDatabase(UserEmail, UserPassword)
-		if err != nil {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		cookie := s.createCookie(authedUser.UserID)
-		if cookie != nil {
-			http.SetCookie(w, cookie)
-		} else {
-			log.Println(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, authedUser)
-	}
-}
-
-// handleLdapLogin attempts to authenticate the user by looking up and authenticating
-// via ldap, and then creates the user if not existing and logs them in
-func (s *server) handleLdapLogin() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-		UserEmail := keyVal["warriorEmail"].(string)
-		UserPassword := keyVal["warriorPassword"].(string)
-
-		authedUser, err := s.authAndCreateUserLdap(UserEmail, UserPassword)
-		if err != nil {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		cookie := s.createCookie(authedUser.UserID)
-		if cookie != nil {
-			http.SetCookie(w, cookie)
-		} else {
-			log.Println(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		respondWithJSON(w, http.StatusOK, authedUser)
-	}
-}
-
-// handleLogout clears the user cookie(s) ending session
-func (s *server) handleLogout() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		s.clearUserCookies(w)
-		return
-	}
-}
-
-// handleUserRecruit registers a user as a guest user
-func (s *server) handleUserRecruit() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		AllowGuests := viper.GetBool("config.allow_guests")
-		if !AllowGuests {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		keyVal := getJSONRequestBody(r, w)
-
-		UserName := keyVal["warriorName"].(string)
-
-		newUser, err := s.database.CreateUserGuest(UserName)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		s.createUserCookie(w, false, newUser.UserID)
-
-		respondWithJSON(w, http.StatusOK, newUser)
-	}
-}
-
-// handleUserEnlist registers a new authenticated user
-func (s *server) handleUserEnlist() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		AllowRegistration := viper.GetBool("config.allow_registration")
-		if !AllowRegistration {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		keyVal := getJSONRequestBody(r, w)
-
-		ActiveUserID, _ := s.validateUserCookie(w, r)
-
-		UserName, UserEmail, UserPassword, accountErr := ValidateUserAccount(
-			keyVal["warriorName"].(string),
-			keyVal["warriorEmail"].(string),
-			keyVal["warriorPassword1"].(string),
-			keyVal["warriorPassword2"].(string),
-		)
-
-		if accountErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		newUser, VerifyID, err := s.database.CreateUserRegistered(UserName, UserEmail, UserPassword, ActiveUserID)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		s.createUserCookie(w, true, newUser.UserID)
-
-		s.email.SendWelcome(UserName, UserEmail, VerifyID)
-
-		respondWithJSON(w, http.StatusOK, newUser)
-	}
-}
-
-// handleForgotPassword attempts to send a password reset email
-func (s *server) handleForgotPassword() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-		UserEmail := keyVal["warriorEmail"].(string)
-
-		ResetID, UserName, resetErr := s.database.UserResetRequest(UserEmail)
-		if resetErr == nil {
-			s.email.SendForgotPassword(UserName, UserEmail, ResetID)
-		}
-
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-}
-
-// handleResetPassword attempts to reset a users password
-func (s *server) handleResetPassword() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-		ResetID := keyVal["resetId"].(string)
-
-		UserPassword, passwordErr := ValidateUserPassword(
-			keyVal["warriorPassword1"].(string),
-			keyVal["warriorPassword2"].(string),
-		)
-
-		if passwordErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		UserName, UserEmail, resetErr := s.database.UserResetPassword(ResetID, UserPassword)
-		if resetErr != nil {
-			log.Println("error attempting to reset user password : " + resetErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		s.email.SendPasswordReset(UserName, UserEmail)
-
-		return
-	}
-}
-
-/*
-	User Handlers
-*/
-
-// handleUpdatePassword attempts to update a users password
-func (s *server) handleUpdatePassword() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-		UserID := r.Context().Value(contextKeyUserID).(string)
-
-		UserPassword, passwordErr := ValidateUserPassword(
-			keyVal["warriorPassword1"].(string),
-			keyVal["warriorPassword2"].(string),
-		)
-
-		if passwordErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		UserName, UserEmail, updateErr := s.database.UserUpdatePassword(UserID, UserPassword)
-		if updateErr != nil {
-			log.Println("error attempting to update user password : " + updateErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		s.email.SendPasswordUpdate(UserName, UserEmail)
-
-		return
-	}
-}
-
-// handleUserProfile returns the users profile if it matches their session
-func (s *server) handleUserProfile() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		User, UserErr := s.database.GetUser(UserID)
-		if UserErr != nil {
-			log.Println("error finding user : " + UserErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, User)
-	}
-}
-
-// handleUserProfileUpdate attempts to update users profile
-func (s *server) handleUserProfileUpdate() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		keyVal := getJSONRequestBody(r, w)
-		UserName := keyVal["warriorName"].(string)
-		UserAvatar := keyVal["warriorAvatar"].(string)
-		NotificationsEnabled, _ := keyVal["notificationsEnabled"].(bool)
-
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		updateErr := s.database.UpdateUserProfile(UserID, UserName, UserAvatar, NotificationsEnabled)
-		if updateErr != nil {
-			log.Println("error attempting to update user profile : " + updateErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		user, UserErr := s.database.GetUser(UserID)
-		if UserErr != nil {
-			log.Println("error reloading user after update : " + UserErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, user)
-	}
-}
-
-// handleAccountVerification attempts to verify a users account
-func (s *server) handleAccountVerification() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-		VerifyID := keyVal["verifyId"].(string)
-
-		verifyErr := s.database.VerifyUserAccount(VerifyID)
-		if verifyErr != nil {
-			log.Println("error attempting to verify user account : " + verifyErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleUserDelete attempts to delete a users account
-func (s *server) handleUserDelete() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		updateErr := s.database.DeleteUser(UserID)
-		if updateErr != nil {
-			log.Println("error attempting to delete user : " + updateErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		s.clearUserCookies(w)
-
-		return
-	}
-}
-
-// handleUserAvatar creates an avatar for the given user by ID
-func (s *server) handleUserAvatar() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-
-		Width, _ := strconv.Atoi(vars["width"])
-		UserID := vars["id"]
-		AvatarGender := govatar.MALE
-		userGender, ok := vars["avatar"]
-		if ok {
-			if userGender == "female" {
-				AvatarGender = govatar.FEMALE
-			}
-		}
-
-		var avatar image.Image
-		if s.config.AvatarService == "govatar" {
-			avatar, _ = govatar.GenerateForUsername(AvatarGender, UserID)
-		} else { // must be goadorable
-			var err error
-			avatar, _, err = image.Decode(bytes.NewReader(adorable.PseudoRandom([]byte(UserID))))
-			if err != nil {
-				log.Fatalln(err)
-			}
-		}
-
-		img := transform.Resize(avatar, Width, Width, transform.Linear)
-		buffer := new(bytes.Buffer)
-
-		if err := png.Encode(buffer, img); err != nil {
-			log.Println("unable to encode image.")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		w.Header().Set("Content-Type", "image/png")
-		w.Header().Set("Content-Length", strconv.Itoa(len(buffer.Bytes())))
-
-		if _, err := w.Write(buffer.Bytes()); err != nil {
-			log.Println("unable to write image.")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-	}
-}
-
-/*
-	API Key Handlers
-*/
-
-// handleAPIKeyGenerate handles generating an API key for a user
-func (s *server) handleAPIKeyGenerate() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		keyVal := getJSONRequestBody(r, w)
-		APIKeyName := keyVal["name"].(string)
-
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		APIKey, keyErr := s.database.GenerateAPIKey(UserID, APIKeyName)
-		if keyErr != nil {
-			log.Println("error attempting to generate api key : " + keyErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, APIKey)
-	}
-}
-
-// handleUserAPIKeys handles getting user API keys
-func (s *server) handleUserAPIKeys() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-
-		APIKeys, keysErr := s.database.GetUserAPIKeys(UserID)
-		if keysErr != nil {
-			log.Println("error retrieving api keys : " + keysErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, APIKeys)
-	}
-}
-
-// handleUserAPIKeyUpdate handles getting user API keys
-func (s *server) handleUserAPIKeyUpdate() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-		APK := vars["keyID"]
-		keyVal := getJSONRequestBody(r, w)
-		active := keyVal["active"].(bool)
-
-		APIKeys, keysErr := s.database.UpdateUserAPIKey(UserID, APK, active)
-		if keysErr != nil {
-			log.Println("error updating api key : " + keysErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, APIKeys)
-	}
-}
-
-// handleUserAPIKeyDelete handles getting user API keys
-func (s *server) handleUserAPIKeyDelete() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-
-		UserID := vars["id"]
-		UserCookieID := r.Context().Value(contextKeyUserID).(string)
-		if UserID != UserCookieID {
-			w.WriteHeader(http.StatusUnauthorized)
-			return
-		}
-		APK := vars["keyID"]
-
-		APIKeys, keysErr := s.database.DeleteUserAPIKey(UserID, APK)
-		if keysErr != nil {
-			log.Println("error deleting api key : " + keysErr.Error() + "\n")
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, APIKeys)
-	}
-}
-
-/*
 	Battle Handlers
 */
 // handleBattleCreate handles creating a battle (arena)
@@ -801,7 +260,7 @@ func (s *server) handleBattleCreate() http.HandlerFunc {
 			return
 		}
 
-		respondWithJSON(w, http.StatusOK, newBattle)
+		s.respondWithJSON(w, http.StatusOK, newBattle)
 	}
 }
 
@@ -816,537 +275,6 @@ func (s *server) handleBattlesGet() http.HandlerFunc {
 			return
 		}
 
-		respondWithJSON(w, http.StatusOK, battles)
-	}
-}
-
-/*
-	Admin Handlers
-*/
-
-// handleAppStats gets the applications stats
-func (s *server) handleAppStats() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		AppStats, err := s.database.GetAppStats()
-
-		if err != nil {
-			http.NotFound(w, r)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, AppStats)
-	}
-}
-
-// handleGetRegisteredUsers gets a list of registered users
-func (s *server) handleGetRegisteredUsers() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Users := s.database.GetRegisteredUsers(Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Users)
-	}
-}
-
-// handleUserCreate registers a new authenticated user
-func (s *server) handleUserCreate() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-
-		UserName, UserEmail, UserPassword, accountErr := ValidateUserAccount(
-			keyVal["warriorName"].(string),
-			keyVal["warriorEmail"].(string),
-			keyVal["warriorPassword1"].(string),
-			keyVal["warriorPassword2"].(string),
-		)
-
-		if accountErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		newUser, VerifyID, err := s.database.CreateUserRegistered(UserName, UserEmail, UserPassword, "")
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		s.email.SendWelcome(UserName, UserEmail, VerifyID)
-
-		respondWithJSON(w, http.StatusOK, newUser)
-	}
-}
-
-// handleUserPromote handles promoting a user to admin
-func (s *server) handleUserPromote() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-
-		err := s.database.PromoteUser(keyVal["warriorId"].(string))
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleUserDemote handles demoting a user to registered
-func (s *server) handleUserDemote() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-
-		err := s.database.DemoteUser(keyVal["warriorId"].(string))
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleCleanBattles handles cleaning up old battles (ADMIN Manaually Triggered)
-func (s *server) handleCleanBattles() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		DaysOld := viper.GetInt("config.cleanup_battles_days_old")
-
-		err := s.database.CleanBattles(DaysOld)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleCleanGuests handles cleaning up old guests (ADMIN Manaually Triggered)
-func (s *server) handleCleanGuests() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		DaysOld := viper.GetInt("config.cleanup_guests_days_old")
-
-		err := s.database.CleanGuests(DaysOld)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleGetOrganizations gets a list of organizations
-func (s *server) handleGetOrganizations() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Organizations := s.database.OrganizationList(Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Organizations)
-	}
-}
-
-// handleGetTeams gets a list of teams
-func (s *server) handleGetTeams() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.TeamList(Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-/*
- * Organization Handlers
- */
-
-// handleGetOrganizationsByUser gets a list of organizations the user is apart of
-func (s *server) handleGetOrganizationsByUser() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Organizations := s.database.OrganizationListByUser(UserID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Organizations)
-	}
-}
-
-// handleGetOrganizationByUser gets an organization with user role
-func (s *server) handleGetOrganizationByUser() http.HandlerFunc {
-	type OrganizationResponse struct {
-		Organization *database.Organization `json:"organization"`
-		Role         string                 `json:"role"`
-	}
-	return func(w http.ResponseWriter, r *http.Request) {
-		UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-
-		Organization, role, err := s.database.OrganizationWithRole(UserID, vars["orgId"])
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, &OrganizationResponse{
-			Organization: Organization,
-			Role:         role,
-		})
-	}
-}
-
-// handleCreateOrganization handles creating an organization with current user as admin
-func (s *server) handleCreateOrganization() http.HandlerFunc {
-	type CreateOrgResponse struct {
-		OrganizationID string `json:"id"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		UserID := r.Context().Value(contextKeyUserID).(string)
-		keyVal := getJSONRequestBody(r, w)
-
-		OrgName := keyVal["name"].(string)
-		OrgId, err := s.database.OrganizationCreate(UserID, OrgName)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		var NewOrg = &CreateOrgResponse{
-			OrganizationID: OrgId,
-		}
-
-		respondWithJSON(w, http.StatusOK, NewOrg)
-	}
-}
-
-// handleGetOrganizationDepartments gets a list of departments associated to the organization
-func (s *server) handleGetOrganizationDepartments() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		OrgID := vars["orgId"]
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.OrganizationDepartmentList(OrgID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-// handleGetDepartmentByUser gets an department with user role
-func (s *server) handleGetDepartmentByUser() http.HandlerFunc {
-	type DepartmentResponse struct {
-		Organization     *database.Organization `json:"organization"`
-		Department       *database.Department   `json:"department"`
-		OrganizationRole string                 `json:"organizationRole"`
-		DepartmentRole   string                 `json:"departmentRole"`
-	}
-	return func(w http.ResponseWriter, r *http.Request) {
-		UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-
-		Organization, OrganizationRole, err := s.database.OrganizationWithRole(UserID, vars["orgId"])
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		Department, DepartmentRole, err := s.database.DepartmentWithRole(UserID, vars["departmentId"])
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		respondWithJSON(w, http.StatusOK, &DepartmentResponse{
-			Organization:     Organization,
-			Department:       Department,
-			OrganizationRole: OrganizationRole,
-			DepartmentRole:   DepartmentRole,
-		})
-	}
-}
-
-// handleCreateDepartment handles creating an organization department
-func (s *server) handleCreateDepartment() http.HandlerFunc {
-	type CreateDepartmentResponse struct {
-		DepartmentID string `json:"id"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		// UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-		keyVal := getJSONRequestBody(r, w)
-
-		OrgName := keyVal["name"].(string)
-		OrgID := vars["orgId"]
-		DepartmentID, err := s.database.DepartmentCreate(OrgID, OrgName)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		var NewDepartment = &CreateDepartmentResponse{
-			DepartmentID: DepartmentID,
-		}
-
-		respondWithJSON(w, http.StatusOK, NewDepartment)
-	}
-}
-
-// handleGetOrganizationTeams gets a list of teams associated to the organization
-func (s *server) handleGetOrganizationTeams() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		OrgID := vars["orgId"]
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.OrganizationTeamList(OrgID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-// handleGetOrganizationUsers gets a list of users associated to the organization
-func (s *server) handleGetOrganizationUsers() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		OrgID := vars["orgId"]
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.OrganizationUserList(OrgID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-// handleGetDepartmentTeams gets a list of teams associated to the department
-func (s *server) handleGetDepartmentTeams() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		DepartmentID := vars["departmentId"]
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.DepartmentTeamList(DepartmentID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-// handleGetDepartmentUsers gets a list of users associated to the department
-func (s *server) handleGetDepartmentUsers() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		DepartmentID := vars["departmentId"]
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.DepartmentUserList(DepartmentID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-// handleCreateOrganizationTeam handles creating an organization team
-func (s *server) handleCreateOrganizationTeam() http.HandlerFunc {
-	type CreateTeamResponse struct {
-		TeamID string `json:"id"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		// UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-		keyVal := getJSONRequestBody(r, w)
-
-		TeamName := keyVal["name"].(string)
-		OrgID := vars["orgId"]
-		TeamID, err := s.database.OrganizationTeamCreate(OrgID, TeamName)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		var NewTeam = &CreateTeamResponse{
-			TeamID: TeamID,
-		}
-
-		respondWithJSON(w, http.StatusOK, NewTeam)
-	}
-}
-
-// handleCreateDepartmentTeam handles creating an department team
-func (s *server) handleCreateDepartmentTeam() http.HandlerFunc {
-	type CreateTeamResponse struct {
-		TeamID string `json:"id"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		// UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-		keyVal := getJSONRequestBody(r, w)
-
-		TeamName := keyVal["name"].(string)
-		DepartmentID := vars["departmentId"]
-		TeamID, err := s.database.DepartmentTeamCreate(DepartmentID, TeamName)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		var NewTeam = &CreateTeamResponse{
-			TeamID: TeamID,
-		}
-
-		respondWithJSON(w, http.StatusOK, NewTeam)
-	}
-}
-
-/*
- * Team Handlers
- */
-
-// handleGetTeamsByUser gets a list of teams the user is apart of
-func (s *server) handleGetTeamsByUser() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		UserID := r.Context().Value(contextKeyUserID).(string)
-		vars := mux.Vars(r)
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Organizations := s.database.TeamListByUser(UserID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Organizations)
-	}
-}
-
-// handleGetTeamUsers gets a list of users associated to the team
-func (s *server) handleGetTeamUsers() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		vars := mux.Vars(r)
-		TeamID := vars["teamId"]
-		Limit, _ := strconv.Atoi(vars["limit"])
-		Offset, _ := strconv.Atoi(vars["offset"])
-
-		Teams := s.database.TeamUserList(TeamID, Limit, Offset)
-
-		respondWithJSON(w, http.StatusOK, Teams)
-	}
-}
-
-// handleCreateTeam handles creating an team with current user as admin
-func (s *server) handleCreateTeam() http.HandlerFunc {
-	type CreateTeamResponse struct {
-		TeamID string `json:"id"`
-	}
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		UserID := r.Context().Value(contextKeyUserID).(string)
-		keyVal := getJSONRequestBody(r, w)
-
-		TeamName := keyVal["name"].(string)
-		TeamID, err := s.database.TeamCreate(UserID, TeamName)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		var NewTeam = &CreateTeamResponse{
-			TeamID: TeamID,
-		}
-
-		respondWithJSON(w, http.StatusOK, NewTeam)
-	}
-}
-
-// handleCreateTeam handles adding user to a team
-func (s *server) handleTeamAddUser() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-
-		vars := mux.Vars(r)
-		TeamID := vars["teamId"]
-		UserEmail := keyVal["email"].(string)
-		Role := keyVal["role"].(string)
-
-		User, UserErr := s.database.GetUserByEmail(UserEmail)
-		if UserErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		_, err := s.database.TeamAddUser(TeamID, User.UserID, Role)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleOrganizationAddUser handles adding user to an organization
-func (s *server) handleOrganizationAddUser() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-
-		vars := mux.Vars(r)
-		OrgID := vars["orgId"]
-		UserEmail := keyVal["email"].(string)
-		Role := keyVal["role"].(string)
-
-		User, UserErr := s.database.GetUserByEmail(UserEmail)
-		if UserErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		_, err := s.database.OrganizationAddUser(OrgID, User.UserID, Role)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-}
-
-// handleDepartmentAddUser handles adding user to an organization department
-func (s *server) handleDepartmentAddUser() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		keyVal := getJSONRequestBody(r, w)
-
-		vars := mux.Vars(r)
-		DepartmentId := vars["departmentId"]
-		UserEmail := keyVal["email"].(string)
-		Role := keyVal["role"].(string)
-
-		User, UserErr := s.database.GetUserByEmail(UserEmail)
-		if UserErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		_, err := s.database.DepartmentAddUser(DepartmentId, User.UserID, Role)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
+		s.respondWithJSON(w, http.StatusOK, battles)
 	}
 }
